requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
babel==2.18.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
courlan==1.3.2
cryptography==46.0.7
dateparser==1.4.0
Deprecated==1.3.1
faiss-cpu==1.13.2
fastapi==0.136.0
filetype==1.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.14.0
htmldate==1.9.4
httpcore==0.17.3
httplib2==0.31.2
httptools==0.7.1
httpx==0.24.1
idna==3.13
jusText==3.0.2
limits==5.8.0
lxml==6.1.0
lxml_html_clean==0.4.4
numpy==2.4.4
packaging==26.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
PyMuPDF==1.27.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
tiktoken==0.12.0
tld==0.13.2
tqdm==4.67.3
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
